refactor: Update file paths and improve footer navigation links
