+++ src/app/[locale]/[...rest]/page.tsx
