+++ src/app/[locale]/restaurants/page.tsx
