+++ src/app/not-found.tsx
@@ -1,5 +1,13 @@
-import NotFound from "@/components/not-found";
+"use client";
 
-export default function NotFoundPage() {
-  return <NotFound />;
+import Error from "next/error";
+
+export default function NotFound() {
+  return (
+    <html lang="en">
+      <body>
+        <Error statusCode={404} />
+      </body>
+    </html>
+  );
 }
